20131129-modify some elements in elements.xml

+++ org.cs2c.vcenter/src/org/cs2c/vcenter/metadata/MetaManager.java
@@ -190,13 +190,13 @@ public class MetaManager {
 					|| (decreaseBlank(recuroneelement.getAttribute("min")) == null))
 				tmpParaMeta.setMin(-2);
 			else
-				tmpParaMeta.setMin(Long.parseLong(decreaseBlank(recuroneelement
+				tmpParaMeta.setMin(Integer.parseInt(decreaseBlank(recuroneelement
 						.getAttribute("min"))));
 			if ((decreaseBlank(recuroneelement.getAttribute("max")) == "")
 					|| (decreaseBlank(recuroneelement.getAttribute("max")) == null))
 				tmpParaMeta.setMax(-2);
 			else
-				tmpParaMeta.setMax(Long.parseLong(decreaseBlank(recuroneelement
+				tmpParaMeta.setMax(Integer.parseInt(decreaseBlank(recuroneelement
 						.getAttribute("max"))));
 
 			tmpParaMeta.setTips(recuroneelement.getAttribute("tips"));
